Refactor devlog commands and update dependencies
